+++ jsqlstreamstore-core/src/main/java/jsqlstreamstore/streams/ErrorMessages.java
@@ -6,11 +6,11 @@ public class ErrorMessages {
         // statics only
     }
 
-    public static String appendFailedWrongExpectedVersion(String streamId, int expectedVersion) {
+    public static String appendFailedWrongExpectedVersion(String streamId, long expectedVersion) {
         return String.format("Append failed due to WrongExpectedVersion.Stream: %s, Expected version: %s", streamId, expectedVersion);
     }
 
-    public static String deleteStreamFailedWrongExpectedVersion(String streamId, int expectedVersion) {
+    public static String deleteStreamFailedWrongExpectedVersion(String streamId, long expectedVersion) {
         return String.format("Delete stream failed due to WrongExpectedVersion.Stream: %s, Expected version: %s", streamId, expectedVersion);
     }
