+++ jsqlstreamstore-core/src/main/java/jsqlstreamstore/streams/MetadataMessage.java
@@ -2,12 +2,14 @@ package jsqlstreamstore.streams;
 
 import com.google.common.base.MoreObjects;
 
+import java.math.BigInteger;
+
 /**
  * From SqlStreamStore MetadataMessage is retrieved from stream jsqlstreamstore.store (db) and then passed to
  * StreamMetadataResult (https://github.com/damianh/SqlStreamStore/blob/ebbb973959729c98836b5f95c3b55ea60eb6ef7a/src/SqlStreamStore.MsSql/MsSqlStreamStore.StreamMetadata.cs)
- * 
+ *
  * This vs StreamMetadata which is from eventstore
- * 
+ *
  *
  */
 public class MetadataMessage {
@@ -16,25 +18,25 @@ public class MetadataMessage {
      * The stream id
      */
 	private String streamId;
-	
+
 	/**
 	 * The max age of messages retained in the stream
 	 */
 	private Integer maxAge;
-	
+
 	/**
 	 * The max count of messages retained in the stream
 	 */
-	private Integer maxCount;
-	
+	private Long maxCount;
+
 	/**
 	 * Custom json
 	 */
 	private String metaJson;
-	
-	public MetadataMessage() {} 
-	
-	public MetadataMessage(String streamId, Integer maxAge, Integer maxCount, String metaJson) {
+
+	public MetadataMessage() {}
+
+	public MetadataMessage(String streamId, Integer maxAge, Long maxCount, String metaJson) {
 	    this.streamId = streamId;
 	    this.maxAge = maxAge;
 	    this.maxCount = maxCount;
@@ -57,11 +59,11 @@ public class MetadataMessage {
         this.maxAge = maxAge;
     }
 
-    public Integer getMaxCount() {
+    public Long getMaxCount() {
         return maxCount;
     }
 
-    public void setMaxCount(Integer maxCount) {
+    public void setMaxCount(Long maxCount) {
         this.maxCount = maxCount;
     }
 
@@ -72,7 +74,7 @@ public class MetadataMessage {
     public void setMetaJson(String metaJson) {
         this.metaJson = metaJson;
     }
-	
+
 	@Override
 	public String toString() {
 	    return MoreObjects.toStringHelper(this)
