+++ jsqlstreamstore-core/src/main/java/jsqlstreamstore/subscriptions/StreamSubscription.java
@@ -9,7 +9,7 @@ public interface StreamSubscription extends Closeable {
 
     String getStreamId();
 
-    Integer getLastVersion();
+    Long getLastVersion();
 
     // Task started();
